+++ acfml.php
@@ -653,9 +653,9 @@ class ACF_Multilingual {
     
     if( $wp_query = $this->resolve_url($url) ) {
       $wp_object = $wp_query->get_queried_object();
+      
       if( $wp_object instanceof \WP_Post ) {
         $new_url = $this->post_types_controller->get_post_link($wp_object, $requested_language);
-        $new_url = $this->add_rewrite_endpoints_to_url($new_url, $wp_query);
         return $new_url;
       } elseif( $wp_object instanceof \WP_Post_Type ) {
         $new_url = $this->post_types_controller->get_post_type_archive_link($wp_object->name, $requested_language);
@@ -665,22 +665,6 @@ class ACF_Multilingual {
     
     // if nothing special was found, only inject the language code
     return $this->simple_convert_url($url, $requested_language);
-  }
-
-  private function add_rewrite_endpoints_to_url($url, $wp_query) {
-    global $wp_rewrite;
-    $endpoints = $wp_rewrite->endpoints ?? [];
-    foreach( $endpoints as $endpoint ) {
-      $endpoint_name = $endpoint[1];
-      $endpoint_query_var = $endpoint[2];
-      if( isset($wp_query->query[$endpoint_query_var]) ) {
-        $url = trailingslashit($url) . "$endpoint_name/";
-        if( !empty($wp_query->query[$endpoint_query_var]) ) {
-          $url .= "{$wp_query->query[$endpoint_query_var]}/";
-        }
-      }
-    }
-    return $url;
   }
 
   /**
@@ -980,7 +964,7 @@ class ACF_Multilingual {
     if( !is_front_page() || is_robots() ) return;
 
     $current_language = $this->get_current_language();
-    $user_language = strtolower(substr($_SERVER['HTTP_ACCEPT_LANGUAGE'], 0, 2));    
+    $user_language = strtolower(substr($_SERVER['HTTP_ACCEPT_LANGUAGE'], 0, 2));
 
     if( $_COOKIE['acfml-language'] ?? null ) return;